Refactored screen names for clarity
