+++ App.js
@@ -1,11 +1,11 @@
 import { createBottomTabNavigator } from 'react-navigation';
-import HomeScreen from './HomeTab';
-import SecondScreen from './VideoTab';
+import HomeScreen from './HomeScreen';
+import VideoScreen from './VideoScreen';
 import ThirdScreen from './ThirdScreen';
-
 //This file will serve only for navigation, import the files for pages above then list them below
+
 export default createBottomTabNavigator({
   Home: HomeScreen,
-  Second: SecondScreen,
+  Videos: VideoScreen,
   Third: ThirdScreen
 });
